+++ Fyber/Fyber/ALInneractiveMediationAdapter.m
@@ -10,7 +10,7 @@
 #import <IASDKCore/IASDKCore.h>
 #import "ASAdTracker.h"
 
-#define ADAPTER_VERSION @"8.2.8.0"
+#define ADAPTER_VERSION @"8.3.2.0"
 
 @interface ALInneractiveMediationAdapterGlobalDelegate : NSObject <IAGlobalAdDelegate>
 @end
@@ -85,8 +85,6 @@ static NSMutableDictionary<NSString *, ALInneractiveMediationAdapter *> *ALInner
         
         NSString *appID = [parameters.serverParameters al_stringForKey: @"app_id"];
         [self log: @"Initializing Inneractive SDK with app id: %@...", appID];
-        
-        [IASDKCore sharedInstance].userID = self.sdk.userIdentifier;
         [IASDKCore sharedInstance].mediationType = [[IAMediationMax alloc] init];
         
         // Passing extra info such as creative id supported in 6.15.0+
@@ -173,15 +171,7 @@ static NSMutableDictionary<NSString *, ALInneractiveMediationAdapter *> *ALInner
     [self updateUserInfoWithRequestParameters: parameters];
     
     NSString *signal = [FMPBiddingManager sharedInstance].biddingToken;
-    if ( [signal al_isValidString] )
-    {
-        [delegate didCollectSignal: signal];
-    }
-    else
-    {
-        [self log: @"Failed to collect signal"];
-        [delegate didFailToCollectSignalWithErrorMessage: nil];
-    }
+    [delegate didCollectSignal: signal];
 }
 
 #pragma mark - MAInterstitialAdapter Methods
@@ -421,8 +411,6 @@ static NSMutableDictionary<NSString *, ALInneractiveMediationAdapter *> *ALInner
 
 - (void)updateUserInfoWithRequestParameters:(id<MAAdapterParameters>)requestParameters
 {
-    [IASDKCore sharedInstance].userID = self.sdk.userIdentifier;
-    
     NSNumber *hasUserConsent = [requestParameters hasUserConsent];
     if ( hasUserConsent != nil )
     {
@@ -641,15 +629,7 @@ static NSMutableDictionary<NSString *, ALInneractiveMediationAdapter *> *ALInner
 
 - (void)IAVideoContentController:(nullable IAVideoContentController *)contentController videoInterruptedWithError:(NSError *)error
 {
-    [self.parentAdapter log: @"Rewarded ad failed to display with error: %@", error];
-    
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wdeprecated-declarations"
-    [self.delegate didFailToDisplayRewardedAdWithError: [MAAdapterError errorWithCode: -4205
-                                                                          errorString: @"Ad Display Failed"
-                                                               thirdPartySdkErrorCode: error.code
-                                                            thirdPartySdkErrorMessage: error.localizedDescription]];
-#pragma clang diagnostic pop
+    [self.parentAdapter log: @"Rewarded video is interrupted and buffering with error: %@", error];
 }
 
 - (void)IAVideoCompleted:(nullable IAVideoContentController *)contentController