+++ Moloco/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 3.1.3.0
+* Certified with Moloco SDK 3.1.3.
+
+## 3.1.2.0
+* Certified with Moloco SDK 3.1.2.
+
+## 3.1.1.0
+* Certified with Moloco SDK 3.1.1.
+
+## 3.0.0.1
+* Lower minOS version to iOS 12.0.
+
+## 3.0.0.0
+* Certified with Moloco SDK 3.0.0.
+
+## 2.2.1.0
+* Certified with Moloco SDK 2.2.1.
+* Updated minimum Xcode requirement to 15.0.
+
 ## 2.1.1.1
 * Add support for leaders.