+++ UnityAds/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 4.12.2.0
+* Certified with UnityAds SDK 4.12.2.
+
+## 4.12.1.0
+* Certified with UnityAds SDK 4.12.1.
+
+## 4.12.0.0
+* Certified with UnityAds SDK 4.12.0.
+
+## 4.11.3.2
+* Revert Swift adapter implementation to prevent `swift::swift_abortRetainUnowned` crashes.
+
+## 4.11.3.1
+* Fully re-written in Swift.
+* Updated minimum AppLovinSDK requirement to 12.0.0.
+
+## 4.11.3.0
+* Certified with UnityAds SDK 4.11.3.
+* Updated the minimum required iOS version to 12.0 to match UnityAds SDK. 
+
+## 4.11.0.0
+* Certified with UnityAds SDK 4.11.0.
+* Updated minimum Xcode requirement to 15.0.
+* Remove deprecated callbacks `didStartRewardedAdVideo` and `didCompleteRewardedAdVideo`.
+
 ## 4.10.0.0
 * Certified with UnityAds SDK 4.10.0.